% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDVIS_plot.R
\name{GDVIS_plot}
\alias{GDVIS_plot}
\title{Plot GDVIS plot}
\usage{
GDVIS_plot(
  input_triangle_parameters,
  show.rendering = TRUE,
  show.names = TRUE,
  x_lower = NULL,
  x_upper = NULL,
  y_lower = NULL,
  y_upper = NULL
)
}
\arguments{
\item{input_triangle_parameters}{the path to the RData GDVIS output, which ends with 3D.triangle_parameters.RData}

\item{show.rendering}{option to show the rendering of the plot in an external window, default is TRUE}

\item{show.names}{option to show the names of the groups, default is TRUE}

\item{x_lower}{change axes to make the plots all the same size or make the legend appear outside of the plot (only for 2D plot)}

\item{x_upper}{change axes to make the plots all the same size or make the legend appear outside of the plot (only for 2D plot)}

\item{y_lower}{change axes to make the plots all the same size or make the legend appear outside of the plot (only for 2D plot)}

\item{y_upper}{change axes to make the plots all the same size or make the legend appear outside of the plot (only for 2D plot)}
}
\description{
This function plots a GDVIS visualization based on GDVIS parameters.
}
