% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDVIS_plot_2D.R
\name{GDVIS_plot_2D}
\alias{GDVIS_plot_2D}
\title{Plot GDVIS 2D plot}
\usage{
GDVIS_plot_2D(
  input_triangle_parameters,
  x_lower = NULL,
  x_upper = NULL,
  y_lower = NULL,
  y_upper = NULL
)
}
\arguments{
\item{input_triangle_parameters}{the path to the RData GDVIS output, which ends with 2D.triangle_parameters.RData}

\item{x_lower}{set lower value for the x axis}

\item{x_upper}{set upper value for the x axis}

\item{y_lower}{set lower value for the y asis}

\item{y_upper}{set upper value for the y axis}
}
\description{
This function plots a GDVIS visualization based on GDVIS parameters.
}
