% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDVIS_plot_2D_2D.R
\name{GDVIS_plot_2D_2D}
\alias{GDVIS_plot_2D_2D}
\title{Plot GDVIS 2D.2D plot}
\usage{
GDVIS_plot_2D_2D(
  input_triangle_parameters,
  show.rendering = TRUE,
  webversion = FALSE,
  show.names = TRUE
)
}
\arguments{
\item{input_triangle_parameters}{the path to the RData GDVIS output, which ends with 2D_2D.triangle_parameters.RData}

\item{show.rendering}{option to show the rendering of the plot in an external window, default is TRUE}

\item{webversion}{needs to stay on FALSE}

\item{show.names}{option to show the names of the groups, default is true}
}
\description{
This function plots a GDVIS visualization based on GDVIS parameters.
}
