% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDVIS_plot_3D.R
\name{GDVIS_plot_3D}
\alias{GDVIS_plot_3D}
\title{Plot GDVIS 3D plot}
\usage{
GDVIS_plot_3D(
  input_triangle_parameters,
  show.rendering = TRUE,
  show.names = TRUE,
  webversion = FALSE
)
}
\arguments{
\item{input_triangle_parameters}{the path to the RData GDVIS output, which ends with 3D.triangle_parameters.RData}

\item{show.rendering}{option to show the rendering of the plot in an external window, default is TRUE}

\item{show.names}{option to show the names of the groups, default is TRUE}

\item{webversion}{needs to stay on FALSE}
}
\description{
This function plots a GDVIS visualization based on GDVIS parameters.
}
